Update architecture.md
1) Line 7 changed to: "Deposits/Withdrawals happen within the contract, and it keeps its own accounting records."
In the same line, "On a regular basis (e.g. daily basis)" - is it regular basis or daily basis? If it's daily then we should use the word daily.
In the same line, "," before thus. 
In the same line, "Anyone can call this method." - we mean any user right? shall we change it to "Any user can call this method."

2) Line 10: "(deposits + rewards + other additions)" - other additions seems a bit ambiguous. Shall we give an example here saying "deposits + rewards other additions such as...)" 
In the same line, added "base" before "Token" to keep it with consistent with the subsequent statements.

3) Line 12: "Some protocols may incur in slashing, thus the exchange rate can in turn decrease." This is a bit confusing. It is usually the validators whose deposit gets slashed, right? Shall we not just say, "Some validators' deposit may get slashed, which might affect the exchange rate." Either we should explain this a bit more or remove this altogether. Otherwise, it's a bit ambiguous and creates confusion. 

4) Line 15: changed to "As the staking rewards accrue, ..." 
When we are talking about a percentage of the rewards here, should we not specify the percentage here? 

5) Line 18: changed to, "The contract will stake a significant portion of the funds, about 95%, leaving a certain amount for liquidity in the contract," 
replace "provide" with "support" 
did some language edits in the same line 

6) Lines 21 and 25: language edits

7) Lines 32 and 33: passive voice to active